requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
image==1.5.33
numpy==1.23.4
tensorflow==2.12.0
pydantic==1.10.7
Pillow==9.3.0
python-multipart==0.0.6